Move all game logic into original file
